Integrate pod quota information into /health enndpoint
