+++ binderhub/app.py
@@ -34,7 +34,7 @@ from .main import MainHandler, ParameterizedMainHandler, LegacyRedirectHandler
 from .repoproviders import (GitHubRepoProvider, GitRepoProvider,
                             GitLabRepoProvider, GistRepoProvider,
                             ZenodoProvider)
-from .metrics import MetricsHandler, PodQuotaHandler
+from .metrics import MetricsHandler
 
 from .utils import ByteSpecification, url_path_join
 from .events import EventLog
@@ -620,7 +620,6 @@ class BinderHub(Application):
                 {'path': os.path.join(self.tornado_settings['static_path'], 'images')}),
             (r'/about', AboutHandler),
             (r'/health', HealthHandler, {'hub_url': self.hub_url}),
-            (r'/_pod_quota', PodQuotaHandler),
             (r'/', MainHandler),
             (r'.*', Custom404),
         ]
